Tidied up required changes to get parallel run
Removed quite a few changes, so there are just focussed changes:
This uses its own Description, so we can add and set a class.  Need to
do this because the surefire parallel runner falls back to just finding
the name of the running class in the Thread.
To this end set the class name, and cascaded it through the
Descriptions.
Removed the Scenario completed event firing, a there is no testMethod
for this the reporter falls apart if you tell it the test succeeded.